Fixed Tests & sending caller via uriVariable

+++ device-blueprint/tests/device.test.js
@@ -2,8 +2,8 @@ const axios = require('axios');
 // import shell
 const shell = require('shelljs');
 
-const base_url = 'http://localhost:3000/coffee-machine';
-const base_url_2 = 'http://host.docker.internal:3001/smart-fridge';
+const base_url = 'http://localhost:3010/coffee-machine';
+const base_url_2 = 'http://host.docker.internal:3011/smart-fridge';
 const base_config = {headers: {'Content-Type': 'application/json'}}
 
 // remove the docker container wot_device_1 before running the test, using a shell command
@@ -71,8 +71,16 @@ test('GET subscribe Event & trigger -> [200]', async () => {
 
 
 // Test make request
-test('POST make_request -> [200]', async () => {
+test('POST make_request (property) -> [200]', async () => {
     let url = `${base_url}/actions/make_request?method=GET&url=${base_url_2}/properties/waterLevel`
+    const response = await axios.post(url, {}, base_config);
+    console.log(response.data);
+    expect(response.status).toBe(200);
+});
+
+// Test make request
+test('POST make_request (action) -> [200]', async () => {
+    let url = `${base_url}/actions/make_request?method=POST&url=${base_url_2}/actions/makeIce`
     const response = await axios.post(url, {}, base_config);
     console.log(response.data);
     expect(response.status).toBe(200);
@@ -82,9 +90,13 @@ test('POST make_request -> [200]', async () => {
 
 // after 10 seconds, remove the docker container wot_device_1 after running the test, using a shell command
 
-setTimeout(() => {
-        shell.exec('docker rm -f wot-device-1');
-        shell.exec('docker rm -f wot-device-2');
-    },
-    1000
-)
+let remove_container = process.env.REMOVE_CONTAINER;
+
+if (remove_container === 'true') {
+    setTimeout(() => {
+            shell.exec('docker rm -f wot-device-1');
+            shell.exec('docker rm -f wot-device-2');
+        },
+        10000
+    )
+}
